ver0.8.14

+++ main/java/dqr/blocks/chest/InventoryDqmEnderChest.java
@@ -150,7 +150,8 @@ public class InventoryDqmEnderChest implements IInventory
 		        for(int i = 0; i < tags.tagCount(); i++)
 		        {
 		            NBTTagCompound tagCompound = tags.getCompoundTagAt(i);
-		            int slot = tagCompound.getByte("Slot");
+
+		            int slot = tagCompound.getInteger("Slot");
 		            if(slot >= 0 && slot < items.length)
 		            {
 		                items[slot] = ItemStack.loadItemStackFromNBT(tagCompound);
@@ -174,7 +175,7 @@ public class InventoryDqmEnderChest implements IInventory
             if(items[i] != null)
             {
                 NBTTagCompound compound = new NBTTagCompound();
-                compound.setByte("Slot", (byte)i);
+                compound.setInteger("Slot", i);
                 items[i].writeToNBT(compound);
                 tagList.appendTag(compound);
             }
